refactor: consolidate auth configuration
