+++ packages/cli/src/runtime/loginOptions.ts
@@ -1,5 +1,5 @@
 import { DEFAULT_OAUTH_PROVIDER } from '@auth/config';
-import { isOAuthProvider, type OAuthProvider } from '@auth/sharedConfig';
+import { isOAuthProvider, type OAuthProvider } from '@auth/config';
 import { isNonEmptyString } from '@utils/text/stringUtils';
 
 import { CLI_OAUTH_PROVIDER_INPUTS } from './oauthProviderDisplay';
