+++ src/test-kernel/controllers/SettingsModelSelectionController.vitest.ts
@@ -5,7 +5,7 @@ import {
   SettingsModelSelectionController,
   type SettingsModelSelectionState,
 } from '@controllers/settingsView/SettingsModelSelectionController';
-import { MAX_TIER } from '@auth/sharedConfig';
+import { MAX_TIER } from '@auth/config';
 import { buildBasicModelOptionsData } from '@model/modelOptionsBasic';
 import type { ModelOptionData } from '@shared/schemas';
 import { DEFAULT_HELPER_MODEL } from '@shared/constants/providers';
